vibepalette/requirements.txt
flask
scikit-learn
pillow
numpy
